requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.10.15
chardet==3.0.4
Click==7.0
click-plugins==1.0.4
colorama==0.4.0
cryptography==2.1.4
enum34==1.1.6
idna==2.7
ipaddr==2.2.0
ipaddress==1.0.17
keyring==10.6.0
keyrings.alt==3.0
kombu==4.2.1
pycrypto==2.6.1
pygobject==3.26.1
python-nmap==0.6.1
pytz==2018.7
pyxdg==0.25
redis==2.10.6
requests==2.20.0
SecretStorage==2.3.1
shodan==1.10.4
six==1.11.0
urllib3==1.24.1
vine==1.1.4
XlsxWriter==1.1.2
